refactor(cli): consolidate default argument handling in tests
- Replaced manual default argument assignments with a centralized `withInitDefaults` function in both extension and init tests.
- This change enhances code clarity and maintainability by reducing redundancy in default value assignments across test files.
