+++ packages/cli/test-utils/index.ts
@@ -2,3 +2,4 @@ export * from './cli';
 export * from './CLIDeployTestFixture';
 export * from './fixtures';
 export * from './TestDatabase';
+export * from './init-argv';
